MOB-1704: rename saved searches slice
