+++ src/i18n/l10n/en.ftl.json
@@ -120,6 +120,8 @@
   "CHECK-YOUR-EMAIL": "CHECK YOUR EMAIL!",
   "Choose-a-date": "Choose a date",
   "Choose-a-date-time": "Choose a date & time",
+  "CHOOSE-A-SAVED-SEARCH": "CHOOSE A SAVED SEARCH",
+  "Choose-a-saved-search": "Choose a saved search",
   "Choose-a-time": "Choose a time",
   "CHOOSE-A-USERNAME": "CHOOSE A USERNAME",
   "Chooses-taxon": "Chooses taxon",
@@ -220,6 +222,7 @@
   "Delete-observation": "Delete observation",
   "DELETE-OBSERVATION--question": "DELETE OBSERVATION?",
   "Delete-photo": "Delete photo",
+  "Delete-saved-search": "Delete saved search",
   "Delete-sound": "Delete sound",
   "Deleting-x-of-y--observations": "Deleting { $currentDeleteCount } { $total ->\n  [one] of { $total }\n *[other] of { $total } observations\n}",
   "Deleting-x-of-y-observations-2": "Deleting { $currentDeleteCount } { $total ->\n  [one] observation\n *[other] of { $total } observations\n}",
@@ -686,6 +689,7 @@
   "Ranks-ZOOSUBSECTION": "ZOOSUBSECTION",
   "Ranks-Zoosubsection": "Zoosubsection",
   "Read-more-on-Wikipedia": "Read more on Wikipedia",
+  "Recent-searches": "Recent searches",
   "Record-a-sound": "Record a sound",
   "Record-animal-sounds": "Record animal sounds",
   "RECORD-NEW-SOUND": "RECORD NEW SOUND",
@@ -701,6 +705,7 @@
   "Remove-Photos": "Remove Photos",
   "Remove-project-filter": "Remove project filter",
   "Remove-taxon-filter": "Remove taxon filter",
+  "Remove-this-saved-search": "Remove this saved search",
   "Remove-user-filter": "Remove user filter",
   "Removes-this-observations-taxon": "Removes this observation's taxon",
   "Removes-your-vote-of-agreement": "Removes your vote of agreement",
@@ -728,7 +733,12 @@
   "SAVE-CHANGES": "SAVE CHANGES",
   "SAVE-FOR-LATER": "SAVE FOR LATER",
   "SAVE-LOCATION": "SAVE LOCATION",
+  "Save-this-search": "Save this search",
   "Saved-Observation": "Saved observation, in queue to upload",
+  "Saved-search-limit-reached": "You have saved as many searches as we can keep",
+  "Saved-search-limit-reached-body": "You can keep { $count } saved searches. Remove one to save another.",
+  "Saved-searches": "Saved searches",
+  "SAVED-SEARCHES": "SAVED SEARCHES",
   "Scientific-Name": "Scientific Name",
   "Scientific-Name-Common-Name": "Scientific Name (Common Name)",
   "SEARCH": "SEARCH",
@@ -941,6 +951,7 @@
   "Would-you-like-to-suggest-the-following-identification": "Would you like to suggest the following identification?",
   "x-comments": "{ $count ->\n  [one] { $count } comment\n *[other] { $count } comments\n}",
   "x-failed": "{ $count ->\n  [one] { $count } failed\n *[other] { $count } failed\n}",
+  "X-filters": "{ $count ->\n  [one] { $count } filter\n *[other] { $count } filters\n}",
   "X-FOLLOWERS": "{ $count ->\n  [one] { $count } FOLLOWER\n *[other] { $count } FOLLOWERS\n}",
   "X-Identifications": "{ $count ->\n  [one] { $count } Identification\n *[other] { $count } Identifications\n}",
   "x-identifications": "{ $count ->\n  [one] { $count } identification\n *[other] { $count } identifications\n}",
