+++ src/stores/useStore.js
@@ -4,7 +4,7 @@ import { createJSONStorage, persist } from "zustand/middleware";
 
 import createExploreSlice from "./createExploreSlice";
 import createExploreV2AdvancedSearchSlice from "./createExploreV2AdvancedSearchSlice";
-import createExploreV2RecentSearchesSlice from "./createExploreV2RecentSearchesSlice";
+import createExploreV2SearchesSlice from "./createExploreV2SearchesSlice";
 import createFeatureFlagSlice from "./createFeatureFlagSlice";
 import createFirebaseTraceSlice from "./createFirebaseTraceSlice";
 import createLayoutSlice from "./createLayoutSlice";
@@ -33,7 +33,7 @@ const useStore = create( persist(
   ( ...args ) => {
     // Let's make our slices
     const slices = [
-      createExploreV2RecentSearchesSlice( ...args ),
+      createExploreV2SearchesSlice( ...args ),
       createExploreSlice( ...args ),
       createExploreV2AdvancedSearchSlice( ...args ),
       createFeatureFlagSlice( ...args ),
